Switch uint_fast16_t to uint_fast32_t for indices lists
To avoid being limited to 65000 long lists

+++ Source/CGL_MathUtil.cpp
@@ -119,10 +119,10 @@ namespace CGL_Math
 		}
 	}
 
-	void CGL_MathUtil::ValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices)
+	void CGL_MathUtil::ValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices)
 	{
 		indices.resize(list.size());
-		for (uint_fast16_t i = 0; i < list.size(); i++)
+		for (uint_fast32_t i = 0; i < list.size(); i++)
 		{
 			indices[i] = i;
 		}
@@ -130,10 +130,10 @@ namespace CGL_Math
 		std::sort(indices.begin(), indices.end(), comparison);
 	}
 
-	void CGL_MathUtil::ValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices)
+	void CGL_MathUtil::ValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices)
 	{
 		indices.resize(list.size());
-		for (uint_fast16_t i = 0; i < list.size(); i++)
+		for (uint_fast32_t i = 0; i < list.size(); i++)
 		{
 			indices[i] = i;
 		}
@@ -142,10 +142,10 @@ namespace CGL_Math
 	}
 
 
-	void CGL_MathUtil::AbsValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices)
+	void CGL_MathUtil::AbsValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices)
 	{
 		indices.resize(list.size());
-		for (uint_fast16_t i = 0; i < list.size(); i++)
+		for (uint_fast32_t i = 0; i < list.size(); i++)
 		{
 			indices[i] = i;
 		}
@@ -153,10 +153,10 @@ namespace CGL_Math
 		std::sort(indices.begin(), indices.end(), comparison);
 	}
 
-	void CGL_MathUtil::AbsValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices)
+	void CGL_MathUtil::AbsValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices)
 	{
 		indices.resize(list.size());
-		for (uint_fast16_t i = 0; i < list.size(); i++)
+		for (uint_fast32_t i = 0; i < list.size(); i++)
 		{
 			indices[i] = i;
 		}
