+++ Source/CGL_MathUtil.h
@@ -99,28 +99,28 @@ namespace CGL_Math
 		* 'indices' is intended to be the size of 'list' and is resized to it to be sure.
 		* if 'indices' is smaller, this may cause re-allocation and impact performance
 		*/
-		static void ValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices);
+		static void ValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices);
 
 		/**
 		* Sorts the vector 'indices' according to descending order of the elements values in the vector 'list'
 		* 'indices' is intended to be the size of 'list' and is resized to it to be sure.
 		* if 'indices' is smaller, this may cause re-allocation and impact performance
 		*/
-		static void ValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices);
+		static void ValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices);
 
 		/**
 		* Sorts the vector 'indices' according to ascending order of the element absolute values in the vector 'list'
 		* 'indices' is intended to be the size of 'list' and is resized to it to be sure.
 		* if 'indices' is smaller, this may cause re-allocation and impact performance
 		*/
-		static void AbsValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices);
+		static void AbsValsToSortedIndicesAscending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices);
 
 		/**
 		* Sorts the vector 'indices' according to descending order of the element absolute values in the vector 'list'
 		* 'indices' is intended to be the size of 'list' and is resized to it to be sure.
 		* if 'indices' is smaller, this may cause re-allocation and impact performance
 		*/
-		static void AbsValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast16_t>& indices);
+		static void AbsValsToSortedIndicesDescending(const std::vector<double>& list, std::vector<uint_fast32_t>& indices);
 
 		/**
 		* Combines the supllied 'vectors' using the supplied 'weights and updates the supplied 'combination' vector reference with the resulting vector:
